vitamin-backend/requirements.txt
fastapi
uvicorn
motor==3.3.2
pymongo==4.6.3
beanie==1.25.0
pydantic
python-dotenv
httpx
passlib[bcrypt]
python-multipart
python-jose[cryptography]
beautifulsoup4
google-generativeai
google-genai